skeleton out resolveServiceDefinition
