--- packages/apollo-language-server/src/providers/federation-info/base.ts
@@ -1,12 +1,0 @@
-// These fields must exist for a service to be a proper
-// federated service
-export interface FederationInfo {
-  sdl: string;
-  name: string;
-  url: string;
-}
-
-export interface ApolloFederationInfoProvider {
-  resolveFederationInfo(): Promise<FederationInfo>;
-  // TODO: onSchemaChange
-}
