+++ packages/apollo-language-server/src/providers/schema/endpoint.ts
@@ -13,10 +13,16 @@ import {
 import { Agent as HTTPSAgent } from "https";
 import { fetch } from "apollo-env";
 import { RemoteServiceConfig } from "../../config";
-import { GraphQLSchemaProvider, SchemaChangeUnsubscribeHandler } from "./base";
+import {
+  GraphQLSchemaProvider,
+  SchemaChangeUnsubscribeHandler,
+  FederationInfo
+} from "./base";
 
 export class EndpointSchemaProvider implements GraphQLSchemaProvider {
   private schema?: GraphQLSchema;
+  private info?: FederationInfo;
+
   constructor(private config: Exclude<RemoteServiceConfig, "name">) {}
   async resolveSchema() {
     if (this.schema) return this.schema;
@@ -55,5 +61,48 @@ export class EndpointSchemaProvider implements GraphQLSchemaProvider {
   ): SchemaChangeUnsubscribeHandler {
     throw new Error("Polling of endpoint not implemented yet");
     return () => {};
+  }
+
+  async resolveServiceDefinition() {
+    if (this.info) return this.info;
+    const { skipSSLValidation, url, headers } = this.config;
+    const options: HttpLink.Options = {
+      uri: url,
+      fetch
+    };
+    if (url.startsWith("https:") && skipSSLValidation) {
+      options.fetchOptions = {
+        agent: new HTTPSAgent({ rejectUnauthorized: false })
+      };
+    }
+
+    const getFederationInfoQuery = `
+      query getFederationInfo {
+        _service {
+          sdl
+        }
+      }
+    `;
+
+    const { data, errors } = (await toPromise(
+      linkExecute(createHttpLink(options), {
+        query: parse(getFederationInfoQuery),
+        context: { headers }
+      })
+    )) as ExecutionResult<{ _service: FederationInfo }>;
+
+    if (errors && errors.length) {
+      // XXX better error handling of GraphQL errors
+      throw new Error(errors.map(({ message }: Error) => message).join("\n"));
+    }
+
+    if (!data || !data._service) {
+      throw new Error(
+        "No data received from server when querying for _service."
+      );
+    }
+
+    this.info = data._service;
+    return data._service;
   }
 }
