+++ packages/apollo-language-server/src/providers/schema/file.ts
@@ -85,4 +85,8 @@ export class FileSchemaProvider implements GraphQLSchemaProvider {
     throw new Error("File watching not implemented yet");
     return () => {};
   }
+
+  async resolveServiceDefinition() {
+    return { sdl: "" };
+  }
 }
