+++ packages/apollo/src/commands/service/push.ts
@@ -74,7 +74,7 @@ export default class ServicePush extends ProjectCommand {
                 "No SDL found in response from federated service. This means that the federated service exposed a `__service` field that did not emit errors, but that did not contain a spec-compliant `sdl` field."
               );
 
-            if (!flags.serviceURL && !info.url)
+            if (!flags.serviceURL)
               throw new Error(
                 "No URL found for federated service. Please provide the URL for the gateway to reach the service via the --serviceURL flag"
               );
@@ -94,8 +94,8 @@ export default class ServicePush extends ProjectCommand {
             } = await project.engine.uploadAndComposePartialSchema({
               id: config.name,
               graphVariant: config.tag,
-              name: flags.serviceName || info.name,
-              url: flags.serviceURL || info.url,
+              name: flags.serviceName,
+              url: flags.serviceURL,
               revision:
                 flags.serviceRevision ||
                 (gitContext && gitContext.commit) ||
@@ -106,7 +106,7 @@ export default class ServicePush extends ProjectCommand {
             });
 
             result = {
-              implementingServiceName: flags.serviceName || info.name,
+              implementingServiceName: flags.serviceName,
               hash: compositionConfig && compositionConfig.schemaHash,
               compositionErrors: errors,
               serviceWasCreated,
